# control-plane/internal/apispec/openapi.yaml
openapi: 3.1.0
info:
  title: Hive API
  version: 0.5.0
  description: 'Hive is a Swarm-native platform control plane. This spec is the single

    source of truth for all API contracts. Every request body, response, and

    error shape is defined here.

    '
servers:
- url: /api/v1
  description: Control plane API v1
components:
  schemas:
...
    StatusResponse:
      description: StatusResponse schema.
      type: object
      additionalProperties: false
      required:
      - status
      properties:
        status:
          type: string
          description: Current status.
      example:
        status: ok
...
    IdResponse:
      description: IdResponse schema.
      type: object
      additionalProperties: false
      required:
      - id
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier.
      example:
        id: 550e8400-e29b-41d4-a716-446655440000
...
    Service:
      description: Service schema.
      type: object
      additionalProperties: false
      required:
      - id
      - name
      properties:
        id:
          type: string
          description: Unique identifier.
        name:
          type: string
          description: Display name.
        image:
          type: string
          description: Docker image reference.
        replicas:
          type: integer
          description: Number of replicas.
        running:
          type: integer
          description: Running.
        labels:
          type: object
          additionalProperties:
            type: string
          description: Key-value labels.
    SwarmSecret:
      description: SwarmSecret schema.
      type: object
      additionalProperties: false
      required:
      - id
      - name
      - createdAt
      properties:
        id:
          type: string
          description: Unique identifier.
        name:
          type: string
          description: Display name.
        labels:
          type: object
          additionalProperties:
            type: string
          description: Key-value labels.
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (ISO 8601).
...
    UpdateSecretRequest:
      description: Request to update or rotate a Swarm secret's data.
      type: object
      additionalProperties: false
      required:
      - data
      properties:
        data:
          type: string
          description: New secret payload (base64-encoded by the server).
        labels:
          type: object
          additionalProperties:
            type: string
          description: Key-value labels.
      example:
        data: s3cr3t-new-value
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    bearerAuth:
      description: bearerAuth schema.
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /services:
    get:
      operationId: listServices
      summary: List Swarm services
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Service list
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
      description: Retrieve services
...
  /secrets:
    get:
      operationId: listSecrets
      summary: List Swarm secrets
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Secret list
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SwarmSecret'
      description: Retrieve secrets
    post:
      operationId: createSecret
      summary: Create Swarm secret
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              - data
              properties:
                name:
                  type: string
                data:
                  type: string
                labels:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        '201':
          description: Secret created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
      description: Create secrets
  /secrets/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
      description: Secret ID.
    put:
      operationId: updateSecret
      summary: Update secret
      description: Swarm secrets are immutable, so this removes the secret and creates a replacement with the same name; the replacement has a DIFFERENT secret ID. Any service still referencing the old ID must be updated. Returns 409 if the secret is in use by one or more services.
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSecretRequest'
      responses:
        '200':
          description: Secret updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/BadGateway'
    delete:
      operationId: deleteSecret
      summary: Delete secret
      description: Removes a Swarm secret.
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Secret removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'
  /secrets/{id}/rotate:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
      description: Secret ID.
    post:
      operationId: rotateSecret
      summary: Rotate secret
      description: Rotates the secret's payload. Because Swarm secrets are immutable, rotation creates a versioned successor secret with a NEW secret ID, re-points the services that referenced the old secret, and removes the old secret.
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSecretRequest'
      responses:
        '200':
          description: Secret rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/BadGateway'
  /configs:
    get:
      operationId: listConfigs
      summary: List Swarm configs
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Config list
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SwarmConfig'
      description: Retrieve configs
    post:
      operationId: createConfig
      summary: Create Swarm config
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              - data
              properties:
                name:
                  type: string
                data:
                  type: string
                labels:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        '201':
          description: Config created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
      description: Create configs
  /configs/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
      description: Config ID.
    put:
      operationId: updateConfig
      summary: Update config
      description: Swarm configs are immutable, so this removes the config and creates a replacement with the same name; the replacement has a DIFFERENT config ID. Any service still referencing the old ID must be updated. Returns 409 if the config is in use by one or more services.
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConfigRequest'
      responses:
        '200':
          description: Config updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/BadGateway'
    delete:
      operationId: deleteConfig
      summary: Delete config
      description: Removes a Swarm config.
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Config removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'
  /networks:
    get:
      operationId: listNetworks
      summary: List Swarm networks
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Network list
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Network'
      description: Retrieve networks
    post:
      operationId: createNetwork
      summary: Create Swarm network
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              - driver
              properties:
                name:
                  type: string
                driver:
                  type: string
                  default: overlay
                internal:
                  type: boolean
                  default: false
                attachable:
                  type: boolean
                  default: false
                labels:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        '201':
          description: Network created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
      description: Create networks
  /networks/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
      description: Network ID.
    delete:
      operationId: deleteNetwork
      summary: Delete network
      description: Removes a Swarm network.
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Network removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'
...
  /settings/ssh-keys:
    get:
      operationId: listSshKeys
      summary: List SSH keys
      description: Lists stored SSH key pairs (private key material is never returned).
      security:
      - bearerAuth: []
      responses:
        '200':
          description: SSH key list
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SSHKey'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createSshKey
      summary: Store SSH key
      description: Stores an SSH key pair for later use when provisioning servers.
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSSHKeyRequest'
      responses:
        '201':
          description: SSH key stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...